selenium/requirements.txt
pyyaml
requests
black
flake8
robotframework
robotframework-seleniumlibrary
chromedriver-autoinstaller
pyotp